Nuevo módulo de formularios implementado en las rutas
